Data.Array.Accelerate.Data.Complex

Description

Complex numbers

Rectangular from

data Complex a :: * -> *

Complex numbers are an algebraic type.

For a complex number z, abs z is a number with the magnitude of z, but oriented in the positive real direction, whereas signum z has the phase of z, but unit magnitude.

Constructors

!a :+ !a infix 6
forms a complex number from its real and imaginary rectangular components.

real :: Elt a => Exp (Complex a) -> Exp a

Return the real part of a complex number

imag :: Elt a => Exp (Complex a) -> Exp a

Return the imaginary part of a complex number

Polar form

mkPolar :: forall a. Floating a => Exp a -> Exp a -> Exp (Complex a)

Form a complex number from polar components of magnitude and phase.

cis :: forall a. Floating a => Exp a -> Exp (Complex a)

cis t is a complex value with magnitude 1 and phase t (modulo 2*pi).

polar :: RealFloat a => Exp (Complex a) -> Exp (a, a)

The function polar takes a complex number and returns a (magnitude, phase) pair in canonical form: the magnitude is non-negative, and the phase in the range (-pi, pi]; if the magnitude is zero, then so is the phase.

magnitude :: RealFloat a => Exp (Complex a) -> Exp a

The non-negative magnitude of a complex number

phase :: RealFloat a => Exp (Complex a) -> Exp a

The phase of a complex number, in the range (-pi, pi]. If the magnitude is zero, then so is the phase.

Conjugate

conjugate :: Num a => Exp (Complex a) -> Exp (Complex a)

Return the complex conjugate of a complex number, defined as

conjugate(Z) = X - iY