Data.Array.Accelerate.Data.Maybe

Description

Since: 1.2.0.0

Documentation

data Maybe a

The Maybe type encapsulates an optional value. A value of type Maybe a either contains a value of type a (represented as Just a), or it is empty (represented as Nothing). Using Maybe is a good way to deal with errors or exceptional cases without resorting to drastic measures such as error.

The Maybe type is also a monad. It is a simple kind of error monad, where all errors are represented by Nothing. A richer error monad can be built using the Either type.

Constructors

Nothing

Just a

just :: Elt a => Exp a -> Exp (Maybe a)

Lift a value into a Just constructor

nothing :: forall a. Elt a => Exp (Maybe a)

The Nothing constructor

maybe :: (Elt a, Elt b) => Exp b -> (Exp a -> Exp b) -> Exp (Maybe a) -> Exp b

The maybe function takes a default value, a function, and a Maybe value. If the Maybe value is nothing, the default value is returned; otherwise, it applies the function to the value inside the Just and returns the result

isJust :: Elt a => Exp (Maybe a) -> Exp Bool

Returns True if the argument is of the form Just _

isNothing :: Elt a => Exp (Maybe a) -> Exp Bool

Returns True if the argument is Nothing

fromMaybe :: Elt a => Exp a -> Exp (Maybe a) -> Exp a

The fromMaybe function takes a default value and a Maybe value. If the Maybe is Nothing, the default value is returned; otherwise, it returns the value contained in the Maybe.

fromJust :: Elt a => Exp (Maybe a) -> Exp a

The fromJust function extracts the element out of the Just constructor. If the argument was actually Nothing, you will get an undefined value instead.

justs :: (Shape sh, Slice sh, Elt a) => Acc (Array (sh :. Int) (Maybe a)) -> Acc (Vector a, Array sh Int)

Extract from an array all of the Just values, together with a segment descriptor indicating how many elements along each dimension were returned.

Orphan instances

Functor Maybe

Instance details

Methods

fmap :: (Elt a, Elt b, Elt (Maybe a), Elt (Maybe b)) => (Exp a -> Exp b) -> Exp (Maybe a) -> Exp (Maybe b)

(<$) :: (Elt a, Elt b, Elt (Maybe a), Elt (Maybe b)) => Exp a -> Exp (Maybe b) -> Exp (Maybe a)

Elt a => IsProduct Elt (Maybe a)

Instance details

Associated Types

type ProdRepr (Maybe a) :: Type

Methods

fromProd :: Maybe a -> ProdRepr (Maybe a)

toProd :: ProdRepr (Maybe a) -> Maybe a

prod :: ProdR Elt (ProdRepr (Maybe a))

(Lift Exp a, Elt (Plain a)) => Lift Exp (Maybe a)

Instance details

Associated Types

type Plain (Maybe a) :: Type

Methods

lift :: Maybe a -> Exp (Plain (Maybe a))

(Semigroup (Exp a), Elt a) => Semigroup (Exp (Maybe a))

Instance details

Methods

(<>) :: Exp (Maybe a) -> Exp (Maybe a) -> Exp (Maybe a)

sconcat :: NonEmpty (Exp (Maybe a)) -> Exp (Maybe a)

stimes :: Integral b => b -> Exp (Maybe a) -> Exp (Maybe a)

(Monoid (Exp a), Elt a) => Monoid (Exp (Maybe a))

Instance details

Methods

mempty :: Exp (Maybe a)

mappend :: Exp (Maybe a) -> Exp (Maybe a) -> Exp (Maybe a)

mconcat :: [Exp (Maybe a)] -> Exp (Maybe a)

Elt a => Elt (Maybe a)

Instance details

Associated Types

type EltRepr (Maybe a) :: Type

Methods

eltType :: TupleType (EltRepr (Maybe a))

fromElt :: Maybe a -> EltRepr (Maybe a)

toElt :: EltRepr (Maybe a) -> Maybe a

Eq a => Eq (Maybe a)

Instance details

Methods

(==) :: Exp (Maybe a) -> Exp (Maybe a) -> Exp Bool

(/=) :: Exp (Maybe a) -> Exp (Maybe a) -> Exp Bool

Ord a => Ord (Maybe a)

Instance details

Methods

(<) :: Exp (Maybe a) -> Exp (Maybe a) -> Exp Bool

(>) :: Exp (Maybe a) -> Exp (Maybe a) -> Exp Bool

(<=) :: Exp (Maybe a) -> Exp (Maybe a) -> Exp Bool

(>=) :: Exp (Maybe a) -> Exp (Maybe a) -> Exp Bool

min :: Exp (Maybe a) -> Exp (Maybe a) -> Exp (Maybe a)

max :: Exp (Maybe a) -> Exp (Maybe a) -> Exp (Maybe a)

compare :: Exp (Maybe a) -> Exp (Maybe a) -> Exp Ordering