Data.Array.Accelerate.Data.Ratio

Description

Standard functions on rational numbers

Since: 1.3.0.0

Documentation

data Ratio a

Rational numbers, with numerator and denominator of some Integral type.

Note that Ratio's instances inherit the deficiencies from the type parameter's. For example, Ratio Natural's Num instance has similar problems to Natural's.

(%) :: Integral a => Exp a -> Exp a -> Exp (Ratio a) infixl 7

Form the ratio of two integral numbers

pattern (:%) :: Elt a => Exp a -> Exp a -> Exp (Ratio a)

numerator :: Exp (Ratio a) -> Elt a -> Exp a

denominator :: Exp (Ratio a) -> Elt a -> Exp a

Orphan instances

Elt a => IsProduct Elt (Ratio a)

Instance details

Associated Types

type ProdRepr (Ratio a) :: Type

Methods

fromProd :: Ratio a -> ProdRepr (Ratio a)

toProd :: ProdRepr (Ratio a) -> Ratio a

prod :: ProdR Elt (ProdRepr (Ratio a))

(FromIntegral a b, Integral b) => FromIntegral a (Ratio b)

Instance details

Methods

fromIntegral :: Exp a -> Exp (Ratio b)

Integral a => Enum (Exp (Ratio a))

Instance details

Methods

succ :: Exp (Ratio a) -> Exp (Ratio a)

pred :: Exp (Ratio a) -> Exp (Ratio a)

toEnum :: Int -> Exp (Ratio a)

fromEnum :: Exp (Ratio a) -> Int

enumFrom :: Exp (Ratio a) -> [Exp (Ratio a)]

enumFromThen :: Exp (Ratio a) -> Exp (Ratio a) -> [Exp (Ratio a)]

enumFromTo :: Exp (Ratio a) -> Exp (Ratio a) -> [Exp (Ratio a)]

enumFromThenTo :: Exp (Ratio a) -> Exp (Ratio a) -> Exp (Ratio a) -> [Exp (Ratio a)]

Integral a => Fractional (Exp (Ratio a))

Instance details

Methods

(/) :: Exp (Ratio a) -> Exp (Ratio a) -> Exp (Ratio a)

recip :: Exp (Ratio a) -> Exp (Ratio a)

fromRational :: Rational -> Exp (Ratio a)

Integral a => Num (Exp (Ratio a))

Instance details

Methods

(+) :: Exp (Ratio a) -> Exp (Ratio a) -> Exp (Ratio a)

(-) :: Exp (Ratio a) -> Exp (Ratio a) -> Exp (Ratio a)

(*) :: Exp (Ratio a) -> Exp (Ratio a) -> Exp (Ratio a)

negate :: Exp (Ratio a) -> Exp (Ratio a)

abs :: Exp (Ratio a) -> Exp (Ratio a)

signum :: Exp (Ratio a) -> Exp (Ratio a)

fromInteger :: Integer -> Exp (Ratio a)

(Integral a, FromIntegral a Int64) => RealFrac (Ratio a)

Instance details

Methods

properFraction :: (Integral b, FromIntegral Int64 b) => Exp (Ratio a) -> (Exp b, Exp (Ratio a))

truncate :: (Integral b, FromIntegral Int64 b) => Exp (Ratio a) -> Exp b

round :: (Integral b, FromIntegral Int64 b) => Exp (Ratio a) -> Exp b

ceiling :: (Integral b, FromIntegral Int64 b) => Exp (Ratio a) -> Exp b

floor :: (Integral b, FromIntegral Int64 b) => Exp (Ratio a) -> Exp b

Elt a => Elt (Ratio a)

Instance details

Associated Types

type EltRepr (Ratio a) :: Type

Methods

eltType :: TupleType (EltRepr (Ratio a))

fromElt :: Ratio a -> EltRepr (Ratio a)

toElt :: EltRepr (Ratio a) -> Ratio a

Integral a => Eq (Ratio a)

Instance details

Methods

(==) :: Exp (Ratio a) -> Exp (Ratio a) -> Exp Bool

(/=) :: Exp (Ratio a) -> Exp (Ratio a) -> Exp Bool

Integral a => Ord (Ratio a)

Instance details

Methods

(<) :: Exp (Ratio a) -> Exp (Ratio a) -> Exp Bool

(>) :: Exp (Ratio a) -> Exp (Ratio a) -> Exp Bool

(<=) :: Exp (Ratio a) -> Exp (Ratio a) -> Exp Bool

(>=) :: Exp (Ratio a) -> Exp (Ratio a) -> Exp Bool

min :: Exp (Ratio a) -> Exp (Ratio a) -> Exp (Ratio a)

max :: Exp (Ratio a) -> Exp (Ratio a) -> Exp (Ratio a)

compare :: Exp (Ratio a) -> Exp (Ratio a) -> Exp Ordering

(Integral a, ToFloating a b) => ToFloating (Ratio a) b

Instance details

Methods

toFloating :: Exp (Ratio a) -> Exp b