Data.Array.Accelerate.Interpreter

Description

This interpreter is meant to be a reference implementation of the semantics of the embedded array language. The emphasis is on defining the semantics clearly, not on performance.

Surface types versus representation types

As a general rule, we perform all computations on representation types and we store all data as values of representation types. To guarantee the type safety of the interpreter, this currently implies a lot of conversions between surface and representation types. Optimising the code by eliminating back and forth conversions is fine, but only where it doesn't negatively affects clarity — after all, the main purpose of the interpreter is to serve as an executable specification.

Interpret an array expression

run :: Arrays a => Acc a -> a

Run a complete embedded array program using the reference interpreter.

run1 :: (Arrays a, Arrays b) => (Acc a -> Acc b) -> a -> b

Prepare and run an embedded array program of one argument

streamOut :: Arrays a => Seq [a] -> [a]

Stream a lazily read list of input arrays through the given program, collecting results as we go