Data.Array.Accelerate.Data.Complex

Description

Complex numbers, stored in the usual C-style array-of-struct representation, for easy interoperability.

Rectangular from

data Complex a

Complex numbers are an algebraic type.

For a complex number z, abs z is a number with the magnitude of z, but oriented in the positive real direction, whereas signum z has the phase of z, but unit magnitude.

The Foldable and Traversable instances traverse the real part first.

Note that Complex's instances inherit the deficiencies from the type parameter's. For example, Complex Float's Ord instance has similar problems to Float's.

Constructors

!a :+ !a infix 6
forms a complex number from its real and imaginary rectangular components.

real :: (Elt a, Elt (Complex a)) => Exp (Complex a) -> Exp a

Return the real part of a complex number

imag :: (Elt a, Elt (Complex a)) => Exp (Complex a) -> Exp a

Return the imaginary part of a complex number

Polar form

mkPolar :: forall a. (Floating a, Elt (Complex a)) => Exp a -> Exp a -> Exp (Complex a)

Form a complex number from polar components of magnitude and phase.

cis :: forall a. (Floating a, Elt (Complex a)) => Exp a -> Exp (Complex a)

cis t is a complex value with magnitude 1 and phase t (modulo 2*pi).

polar :: (RealFloat a, Elt (Complex a)) => Exp (Complex a) -> Exp (a, a)

The function polar takes a complex number and returns a (magnitude, phase) pair in canonical form: the magnitude is non-negative, and the phase in the range (-pi, pi]; if the magnitude is zero, then so is the phase.

magnitude :: (RealFloat a, Elt (Complex a)) => Exp (Complex a) -> Exp a

The non-negative magnitude of a complex number

magnitude' :: (RealFloat a, Elt (Complex a)) => Exp (Complex a) -> Exp a

As magnitude, but ignore floating point rounding and use the traditional (simpler to evaluate) definition.

Since: 1.3.0.0

phase :: (RealFloat a, Elt (Complex a)) => Exp (Complex a) -> Exp a

The phase of a complex number, in the range (-pi, pi]. If the magnitude is zero, then so is the phase.

Conjugate

conjugate :: (Num a, Elt (Complex a)) => Exp (Complex a) -> Exp (Complex a)

Return the complex conjugate of a complex number, defined as

conjugate(Z) = X - iY

Orphan instances

Functor Complex
Since: 1.2.0.0

Instance details

Methods

fmap :: (Elt a, Elt b, Elt (Complex a), Elt (Complex b)) => (Exp a -> Exp b) -> Exp (Complex a) -> Exp (Complex b)

(<$) :: (Elt a, Elt b, Elt (Complex a), Elt (Complex b)) => Exp a -> Exp (Complex b) -> Exp (Complex a)

cst a => IsProduct cst (Complex a)

Instance details

Associated Types

type ProdRepr (Complex a) :: Type

Methods

fromProd :: Complex a -> ProdRepr (Complex a)

toProd :: ProdRepr (Complex a) -> Complex a

prod :: ProdR cst (ProdRepr (Complex a))

(Elt a, Elt (Complex a)) => Unlift Exp (Complex (Exp a))

Instance details

Methods

unlift :: Exp (Plain (Complex (Exp a))) -> Complex (Exp a)

(Lift Exp a, Elt (Plain a), Elt (Complex (Plain a))) => Lift Exp (Complex a)

Instance details

Associated Types

type Plain (Complex a) :: Type

Methods

lift :: Complex a -> Exp (Plain (Complex a))

(FromIntegral a b, Num b, Elt (Complex b)) => FromIntegral a (Complex b)

Instance details

Methods

fromIntegral :: Exp a -> Exp (Complex b)

(RealFloat a, Elt (Complex a)) => Floating (Exp (Complex a))

Instance details

Methods

pi :: Exp (Complex a)

exp :: Exp (Complex a) -> Exp (Complex a)

log :: Exp (Complex a) -> Exp (Complex a)

sqrt :: Exp (Complex a) -> Exp (Complex a)

(**) :: Exp (Complex a) -> Exp (Complex a) -> Exp (Complex a)

logBase :: Exp (Complex a) -> Exp (Complex a) -> Exp (Complex a)

sin :: Exp (Complex a) -> Exp (Complex a)

cos :: Exp (Complex a) -> Exp (Complex a)

tan :: Exp (Complex a) -> Exp (Complex a)

asin :: Exp (Complex a) -> Exp (Complex a)

acos :: Exp (Complex a) -> Exp (Complex a)

atan :: Exp (Complex a) -> Exp (Complex a)

sinh :: Exp (Complex a) -> Exp (Complex a)

cosh :: Exp (Complex a) -> Exp (Complex a)

tanh :: Exp (Complex a) -> Exp (Complex a)

asinh :: Exp (Complex a) -> Exp (Complex a)

acosh :: Exp (Complex a) -> Exp (Complex a)

atanh :: Exp (Complex a) -> Exp (Complex a)

log1p :: Exp (Complex a) -> Exp (Complex a)

expm1 :: Exp (Complex a) -> Exp (Complex a)

log1pexp :: Exp (Complex a) -> Exp (Complex a)

log1mexp :: Exp (Complex a) -> Exp (Complex a)

(RealFloat a, Elt (Complex a)) => Fractional (Exp (Complex a))

Instance details

Methods

(/) :: Exp (Complex a) -> Exp (Complex a) -> Exp (Complex a)

recip :: Exp (Complex a) -> Exp (Complex a)

fromRational :: Rational -> Exp (Complex a)

(RealFloat a, Elt (Complex a)) => Num (Exp (Complex a))

Instance details

Methods

(+) :: Exp (Complex a) -> Exp (Complex a) -> Exp (Complex a)

(-) :: Exp (Complex a) -> Exp (Complex a) -> Exp (Complex a)

(*) :: Exp (Complex a) -> Exp (Complex a) -> Exp (Complex a)

negate :: Exp (Complex a) -> Exp (Complex a)

abs :: Exp (Complex a) -> Exp (Complex a)

signum :: Exp (Complex a) -> Exp (Complex a)

fromInteger :: Integer -> Exp (Complex a)

Elt (Complex Double)

Instance details

Associated Types

type EltRepr (Complex Double) :: Type

Methods

eltType :: TupleType (EltRepr (Complex Double))

fromElt :: Complex Double -> EltRepr (Complex Double)

toElt :: EltRepr (Complex Double) -> Complex Double

Elt (Complex Float)

Instance details

Associated Types

type EltRepr (Complex Float) :: Type

Methods

eltType :: TupleType (EltRepr (Complex Float))

fromElt :: Complex Float -> EltRepr (Complex Float)

toElt :: EltRepr (Complex Float) -> Complex Float

Elt (Complex CFloat)

Instance details

Associated Types

type EltRepr (Complex CFloat) :: Type

Methods

eltType :: TupleType (EltRepr (Complex CFloat))

fromElt :: Complex CFloat -> EltRepr (Complex CFloat)

toElt :: EltRepr (Complex CFloat) -> Complex CFloat

Elt (Complex CDouble)

Instance details

Associated Types

type EltRepr (Complex CDouble) :: Type

Methods

eltType :: TupleType (EltRepr (Complex CDouble))

fromElt :: Complex CDouble -> EltRepr (Complex CDouble)

toElt :: EltRepr (Complex CDouble) -> Complex CDouble

Elt (Complex Half)

Instance details

Associated Types

type EltRepr (Complex Half) :: Type

Methods

eltType :: TupleType (EltRepr (Complex Half))

fromElt :: Complex Half -> EltRepr (Complex Half)

toElt :: EltRepr (Complex Half) -> Complex Half

(Eq a, Elt (Complex a)) => Eq (Complex a)

Instance details

Methods

(==) :: Exp (Complex a) -> Exp (Complex a) -> Exp Bool

(/=) :: Exp (Complex a) -> Exp (Complex a) -> Exp Bool