Data.Array.Accelerate.Data.Either

Description

Since: 1.2.0.0

Documentation

data Either a b :: * -> * -> *

The Either type represents values with two possibilities: a value of type Either a b is either Left a or Right b.

The Either type is sometimes used to represent a value which is either correct or an error; by convention, the Left constructor is used to hold an error value and the Right constructor is used to hold a correct value (mnemonic: "right" also means "correct").

Examples

The type Either String Int is the type of values which can be either a String or an Int. The Left constructor can be used only on Strings, and the Right constructor can be used only on Ints:

>>> let s = Left "foo" :: Either String Int
>>> s
Left "foo"
>>> let n = Right 3 :: Either String Int
>>> n
Right 3
>>> :type s
s :: Either String Int
>>> :type n
n :: Either String Int

The fmap from our Functor instance will ignore Left values, but will apply the supplied function to values contained in a Right:

>>> let s = Left "foo" :: Either String Int
>>> let n = Right 3 :: Either String Int
>>> fmap (*2) s
Left "foo"
>>> fmap (*2) n
Right 6

The Monad instance for Either allows us to chain together multiple actions which may fail, and fail overall if any of the individual steps failed. First we'll write a function that can either parse an Int from a Char, or fail.

>>> import Data.Char ( digitToInt, isDigit )
>>> :{
    let parseEither :: Char -> Either String Int
        parseEither c
          | isDigit c = Right (digitToInt c)
          | otherwise = Left "parse error"
>>> :}

The following should work, since both '1' and '2' can be parsed as Ints.

>>> :{
    let parseMultiple :: Either String Int
        parseMultiple = do
          x <- parseEither '1'
          y <- parseEither '2'
          return (x + y)
>>> :}

>>> parseMultiple
Right 3

But the following should fail overall, since the first operation where we attempt to parse 'm' as an Int will fail:

>>> :{
    let parseMultiple :: Either String Int
        parseMultiple = do
          x <- parseEither 'm'
          y <- parseEither '2'
          return (x + y)
>>> :}

>>> parseMultiple
Left "parse error"

Constructors

Left a

Right b

left :: forall a b. (Elt a, Elt b) => Exp a -> Exp (Either a b)

Lift a value into the Left constructor

right :: forall a b. (Elt a, Elt b) => Exp b -> Exp (Either a b)

Lift a value into the Right constructor

either :: (Elt a, Elt b, Elt c) => (Exp a -> Exp c) -> (Exp b -> Exp c) -> Exp (Either a b) -> Exp c

The either function performs case analysis on the Either type. If the value is Left a, apply the first function to a; if it is Right b, apply the second function to b.

isLeft :: (Elt a, Elt b) => Exp (Either a b) -> Exp Bool

Return True if the argument is a Left-value

isRight :: (Elt a, Elt b) => Exp (Either a b) -> Exp Bool

Return True if the argument is a Right-value

fromLeft :: (Elt a, Elt b) => Exp (Either a b) -> Exp a

The fromLeft function extracts the element out of the Left constructor. If the argument was actually Right, you will get an undefined value instead.

fromRight :: (Elt a, Elt b) => Exp (Either a b) -> Exp b

The fromRight function extracts the element out of the Right constructor. If the argument was actually Left, you will get an undefined value instead.

lefts :: (Shape sh, Slice sh, Elt a, Elt b) => Acc (Array (sh :. Int) (Either a b)) -> Acc (Vector a, Array sh Int)

Extract from the array of Either all of the Left elements, together with a segment descriptor indicating how many elements along each dimension were returned.

rights :: (Shape sh, Slice sh, Elt a, Elt b) => Acc (Array (sh :. Int) (Either a b)) -> Acc (Vector b, Array sh Int)

Extract from the array of Either all of the Right elements, together with a segment descriptor indicating how many elements along each dimension were returned.

Orphan instances

(Lift Exp a, Lift Exp b, Elt (Plain a), Elt (Plain b)) => Lift Exp (Either a b)

Associated Types

type Plain (Either a b) :: *

Methods

lift :: Either a b -> Exp (Plain (Either a b))

(Elt a, Elt b) => Semigroup (Exp (Either a b))

Methods

(<>) :: Exp (Either a b) -> Exp (Either a b) -> Exp (Either a b)

sconcat :: NonEmpty (Exp (Either a b)) -> Exp (Either a b)

stimes :: Integral b => b -> Exp (Either a b) -> Exp (Either a b)

Elt a => Functor (Either a)

Methods

fmap :: (Elt a, Elt b, Elt (Either a a), Elt (Either a b)) => (Exp a -> Exp b) -> Exp (Either a a) -> Exp (Either a b)

(<$) :: (Elt a, Elt b, Elt (Either a a), Elt (Either a b)) => Exp a -> Exp (Either a b) -> Exp (Either a a)

(Elt a, Elt b) => Elt (Either a b)

Methods

eltType :: Either a b -> TupleType (EltRepr (Either a b))

fromElt :: Either a b -> EltRepr (Either a b)

toElt :: EltRepr (Either a b) -> Either a b

(Eq a, Eq b) => Eq (Either a b)

Methods

(==) :: Exp (Either a b) -> Exp (Either a b) -> Exp Bool

(/=) :: Exp (Either a b) -> Exp (Either a b) -> Exp Bool

(Ord a, Ord b) => Ord (Either a b)

Methods

(<) :: Exp (Either a b) -> Exp (Either a b) -> Exp Bool

(>) :: Exp (Either a b) -> Exp (Either a b) -> Exp Bool

(<=) :: Exp (Either a b) -> Exp (Either a b) -> Exp Bool

(>=) :: Exp (Either a b) -> Exp (Either a b) -> Exp Bool

min :: Exp (Either a b) -> Exp (Either a b) -> Exp (Either a b)

max :: Exp (Either a b) -> Exp (Either a b) -> Exp (Either a b)

compare :: Exp (Either a b) -> Exp (Either a b) -> Exp Ordering