Data.Array.Accelerate.Data.Functor

Description

A functor performs a uniform action over a parameterised type

This is essentially the same as the standard Haskell Functor class, lifted to Accelarete Exp terms.

Since: 1.2.0.0

Documentation

class Functor f where

The Functor class is used for scalar types which can be mapped over. Instances of Functor should satisfy the following laws:

fmap id      == id
fmap (f . g) == fmap f . fmap g

Minimal complete definition

fmap

Methods

fmap :: (Elt a, Elt b, Elt (f a), Elt (f b)) => (Exp a -> Exp b) -> Exp (f a) -> Exp (f b)

(<$) :: (Elt a, Elt b, Elt (f a), Elt (f b)) => Exp a -> Exp (f b) -> Exp (f a) infixl 4

Replace all locations in the input with the same value. The default definition is fmap . const, but this may be overridden with a more efficient version.

(<$>) :: (Functor f, Elt a, Elt b, Elt (f a), Elt (f b)) => (Exp a -> Exp b) -> Exp (f a) -> Exp (f b) infixl 4

An infix synonym for fmap

The name of this operator is an allusion to $. Note the similarities between their types:

 ($)  ::              (Exp a -> Exp b) -> Exp a     -> Exp b
(<$>) :: Functor f => (Exp a -> Exp b) -> Exp (f a) -> Exp (f b)

Whereas $ is function application, <$> is function application lifted over a Functor.

($>) :: (Functor f, Elt a, Elt b, Elt (f a), Elt (f b)) => Exp (f a) -> Exp b -> Exp (f b) infixl 4

A flipped version of '(<$)'.

void :: (Functor f, Elt a, Elt (f a), Elt (f ())) => Exp (f a) -> Exp (f ())

void value discards or ignores the result of evaluation.