Data.Array.Accelerate.Data.Monoid

Description

Monoid instances for Accelerate

Documentation

class Monoid a where

The class of monoids (types with an associative binary operation that has an identity). Instances should satisfy the following laws:

mappend mempty x = x

mappend x mempty = x

mappend x (mappend y z) = mappend (mappend x y) z

mconcat = foldr mappend mempty

The method names refer to the monoid of lists under concatenation, but there are many other instances.

Some types can be viewed as a monoid in more than one way, e.g. both addition and multiplication on numbers. In such cases we often define newtypes and make those instances of Monoid, e.g. Sum and Product.

Minimal complete definition

mempty, mappend

(<>) :: Monoid m => m -> m -> m infixr 6

An infix synonym for mappend.

Since: 4.5.0.0

newtype Sum a :: * -> *

Monoid under addition.

Constructors

Sum

Fields

getSum :: a

newtype Product a :: * -> *

Monoid under multiplication.

Constructors

Product

Fields

getProduct :: a

Orphan instances

Elt a => Unlift Exp (Sum (Exp a))

Methods

unlift :: Exp (Plain (Sum (Exp a))) -> Sum (Exp a)

Elt a => Unlift Exp (Product (Exp a))

Methods

unlift :: Exp (Plain (Product (Exp a))) -> Product (Exp a)

(Lift Exp a, Elt (Plain a)) => Lift Exp (Sum a)

Associated Types

type Plain (Sum a) :: *

Methods

lift :: Sum a -> Exp (Plain (Sum a))

(Lift Exp a, Elt (Plain a)) => Lift Exp (Product a)

Associated Types

type Plain (Product a) :: *

Methods

lift :: Product a -> Exp (Plain (Product a))

Num a => Num (Exp (Sum a))

Methods

(+) :: Exp (Sum a) -> Exp (Sum a) -> Exp (Sum a)

(-) :: Exp (Sum a) -> Exp (Sum a) -> Exp (Sum a)

(*) :: Exp (Sum a) -> Exp (Sum a) -> Exp (Sum a)

negate :: Exp (Sum a) -> Exp (Sum a)

abs :: Exp (Sum a) -> Exp (Sum a)

signum :: Exp (Sum a) -> Exp (Sum a)

fromInteger :: Integer -> Exp (Sum a)

Num a => Num (Exp (Product a))

Methods

(+) :: Exp (Product a) -> Exp (Product a) -> Exp (Product a)

(-) :: Exp (Product a) -> Exp (Product a) -> Exp (Product a)

(*) :: Exp (Product a) -> Exp (Product a) -> Exp (Product a)

negate :: Exp (Product a) -> Exp (Product a)

abs :: Exp (Product a) -> Exp (Product a)

signum :: Exp (Product a) -> Exp (Product a)

fromInteger :: Integer -> Exp (Product a)

Monoid (Exp ())

Methods

mempty :: Exp ()

mappend :: Exp () -> Exp () -> Exp ()

mconcat :: [Exp ()] -> Exp ()

(Elt a, Elt b, Monoid (Exp a), Monoid (Exp b)) => Monoid (Exp (a, b))

Methods

mempty :: Exp (a, b)

mappend :: Exp (a, b) -> Exp (a, b) -> Exp (a, b)

mconcat :: [Exp (a, b)] -> Exp (a, b)

(Elt a, Elt b, Elt c, Monoid (Exp a), Monoid (Exp b), Monoid (Exp c)) => Monoid (Exp (a, b, c))

Methods

mempty :: Exp (a, b, c)

mappend :: Exp (a, b, c) -> Exp (a, b, c) -> Exp (a, b, c)

mconcat :: [Exp (a, b, c)] -> Exp (a, b, c)

(Elt a, Elt b, Elt c, Elt d, Monoid (Exp a), Monoid (Exp b), Monoid (Exp c), Monoid (Exp d)) => Monoid (Exp (a, b, c, d))

Methods

mempty :: Exp (a, b, c, d)

mappend :: Exp (a, b, c, d) -> Exp (a, b, c, d) -> Exp (a, b, c, d)

mconcat :: [Exp (a, b, c, d)] -> Exp (a, b, c, d)

(Elt a, Elt b, Elt c, Elt d, Elt e, Monoid (Exp a), Monoid (Exp b), Monoid (Exp c), Monoid (Exp d), Monoid (Exp e)) => Monoid (Exp (a, b, c, d, e))

Methods

mempty :: Exp (a, b, c, d, e)

mappend :: Exp (a, b, c, d, e) -> Exp (a, b, c, d, e) -> Exp (a, b, c, d, e)

mconcat :: [Exp (a, b, c, d, e)] -> Exp (a, b, c, d, e)

Num a => Monoid (Exp (Sum a))

Methods

mempty :: Exp (Sum a)

mappend :: Exp (Sum a) -> Exp (Sum a) -> Exp (Sum a)

mconcat :: [Exp (Sum a)] -> Exp (Sum a)

Num a => Monoid (Exp (Product a))

Methods

mempty :: Exp (Product a)

mappend :: Exp (Product a) -> Exp (Product a) -> Exp (Product a)

mconcat :: [Exp (Product a)] -> Exp (Product a)

Elt a => Elt (Sum a)

Methods

eltType :: Sum a -> TupleType (EltRepr (Sum a))

fromElt :: Sum a -> EltRepr (Sum a)

toElt :: EltRepr (Sum a) -> Sum a

Elt a => Elt (Product a)

Methods

eltType :: Product a -> TupleType (EltRepr (Product a))

fromElt :: Product a -> EltRepr (Product a)

toElt :: EltRepr (Product a) -> Product a

Eq a => Eq (Sum a)

Methods

(==*) :: Exp (Sum a) -> Exp (Sum a) -> Exp Bool

(/=*) :: Exp (Sum a) -> Exp (Sum a) -> Exp Bool

Eq a => Eq (Product a)

Methods

(==*) :: Exp (Product a) -> Exp (Product a) -> Exp Bool

(/=*) :: Exp (Product a) -> Exp (Product a) -> Exp Bool

Ord a => Ord (Sum a)

Methods

(<*) :: Exp (Sum a) -> Exp (Sum a) -> Exp Bool

(>*) :: Exp (Sum a) -> Exp (Sum a) -> Exp Bool

(<=*) :: Exp (Sum a) -> Exp (Sum a) -> Exp Bool

(>=*) :: Exp (Sum a) -> Exp (Sum a) -> Exp Bool

min :: Exp (Sum a) -> Exp (Sum a) -> Exp (Sum a)

max :: Exp (Sum a) -> Exp (Sum a) -> Exp (Sum a)

Ord a => Ord (Product a)

Methods

(<*) :: Exp (Product a) -> Exp (Product a) -> Exp Bool

(>*) :: Exp (Product a) -> Exp (Product a) -> Exp Bool

(<=*) :: Exp (Product a) -> Exp (Product a) -> Exp Bool

(>=*) :: Exp (Product a) -> Exp (Product a) -> Exp Bool

min :: Exp (Product a) -> Exp (Product a) -> Exp (Product a)

max :: Exp (Product a) -> Exp (Product a) -> Exp (Product a)