Data.Array.Accelerate.Data.Monoid

Description

Monoid instances for Accelerate

Since: 1.2.0.0

Documentation

class Semigroup a => Monoid a where

The class of monoids (types with an associative binary operation that has an identity). Instances should satisfy the following laws:

x <> mempty = x

mempty <> x = x

x <> (y <> z) = (x <> y) <> z (Semigroup law)

mconcat = foldr '(<>)' mempty

The method names refer to the monoid of lists under concatenation, but there are many other instances.

Some types can be viewed as a monoid in more than one way, e.g. both addition and multiplication on numbers. In such cases we often define newtypes and make those instances of Monoid, e.g. Sum and Product.

NOTE: Semigroup is a superclass of Monoid since base-4.11.0.0.

Minimal complete definition

mempty

Methods

mempty :: a

Identity of mappend

mappend :: a -> a -> a

An associative operation

NOTE: This method is redundant and has the default implementation mappend = '(<>)' since base-4.11.0.0.

mconcat :: [a] -> a

Fold a list using the monoid.

For most types, the default definition for mconcat will be used, but the function is included in the class definition so that an optimized version can be provided for specific types.

(<>) :: Semigroup a => a -> a -> a infixr 6

An associative operation.

newtype Sum a

Monoid under addition.

>>> getSum (Sum 1 <> Sum 2 <> mempty)
3

Constructors

Sum

Fields

getSum :: a

newtype Product a

Monoid under multiplication.

>>> getProduct (Product 3 <> Product 4 <> mempty)
12

Constructors

Product

Fields

getProduct :: a

Orphan instances

Elt a => Unlift Exp (Sum (Exp a))

Instance details

Methods

unlift :: Exp (Plain (Sum (Exp a))) -> Sum (Exp a)

Elt a => Unlift Exp (Product (Exp a))

Instance details

Methods

unlift :: Exp (Plain (Product (Exp a))) -> Product (Exp a)

(Lift Exp a, Elt (Plain a)) => Lift Exp (Sum a)

Instance details

Associated Types

type Plain (Sum a) :: *

Methods

lift :: Sum a -> Exp (Plain (Sum a))

(Lift Exp a, Elt (Plain a)) => Lift Exp (Product a)

Instance details

Associated Types

type Plain (Product a) :: *

Methods

lift :: Product a -> Exp (Plain (Product a))

Bounded a => Bounded (Exp (Sum a))

Instance details

Methods

minBound :: Exp (Sum a)

maxBound :: Exp (Sum a)

Bounded a => Bounded (Exp (Product a))

Instance details

Methods

minBound :: Exp (Product a)

maxBound :: Exp (Product a)

Num a => Num (Exp (Sum a))

Instance details

Methods

(+) :: Exp (Sum a) -> Exp (Sum a) -> Exp (Sum a)

(-) :: Exp (Sum a) -> Exp (Sum a) -> Exp (Sum a)

(*) :: Exp (Sum a) -> Exp (Sum a) -> Exp (Sum a)

negate :: Exp (Sum a) -> Exp (Sum a)

abs :: Exp (Sum a) -> Exp (Sum a)

signum :: Exp (Sum a) -> Exp (Sum a)

fromInteger :: Integer -> Exp (Sum a)

Num a => Num (Exp (Product a))

Instance details

Methods

(+) :: Exp (Product a) -> Exp (Product a) -> Exp (Product a)

(-) :: Exp (Product a) -> Exp (Product a) -> Exp (Product a)

(*) :: Exp (Product a) -> Exp (Product a) -> Exp (Product a)

negate :: Exp (Product a) -> Exp (Product a)

abs :: Exp (Product a) -> Exp (Product a)

signum :: Exp (Product a) -> Exp (Product a)

fromInteger :: Integer -> Exp (Product a)

Num a => Semigroup (Exp (Sum a))
Since: 1.2.0.0

Instance details

Methods

(<>) :: Exp (Sum a) -> Exp (Sum a) -> Exp (Sum a)

sconcat :: NonEmpty (Exp (Sum a)) -> Exp (Sum a)

stimes :: Integral b => b -> Exp (Sum a) -> Exp (Sum a)

Num a => Semigroup (Exp (Product a))
Since: 1.2.0.0

Instance details

Methods

(<>) :: Exp (Product a) -> Exp (Product a) -> Exp (Product a)

sconcat :: NonEmpty (Exp (Product a)) -> Exp (Product a)

stimes :: Integral b => b -> Exp (Product a) -> Exp (Product a)

Monoid (Exp ())

Instance details

Methods

mempty :: Exp ()

mappend :: Exp () -> Exp () -> Exp ()

mconcat :: [Exp ()] -> Exp ()

(Elt a, Elt b, Monoid (Exp a), Monoid (Exp b)) => Monoid (Exp (a, b))

Instance details

Methods

mempty :: Exp (a, b)

mappend :: Exp (a, b) -> Exp (a, b) -> Exp (a, b)

mconcat :: [Exp (a, b)] -> Exp (a, b)

(Elt a, Elt b, Elt c, Monoid (Exp a), Monoid (Exp b), Monoid (Exp c)) => Monoid (Exp (a, b, c))

Instance details

Methods

mempty :: Exp (a, b, c)

mappend :: Exp (a, b, c) -> Exp (a, b, c) -> Exp (a, b, c)

mconcat :: [Exp (a, b, c)] -> Exp (a, b, c)

(Elt a, Elt b, Elt c, Elt d, Monoid (Exp a), Monoid (Exp b), Monoid (Exp c), Monoid (Exp d)) => Monoid (Exp (a, b, c, d))

Instance details

Methods

mempty :: Exp (a, b, c, d)

mappend :: Exp (a, b, c, d) -> Exp (a, b, c, d) -> Exp (a, b, c, d)

mconcat :: [Exp (a, b, c, d)] -> Exp (a, b, c, d)

(Elt a, Elt b, Elt c, Elt d, Elt e, Monoid (Exp a), Monoid (Exp b), Monoid (Exp c), Monoid (Exp d), Monoid (Exp e)) => Monoid (Exp (a, b, c, d, e))

Instance details

Methods

mempty :: Exp (a, b, c, d, e)

mappend :: Exp (a, b, c, d, e) -> Exp (a, b, c, d, e) -> Exp (a, b, c, d, e)

mconcat :: [Exp (a, b, c, d, e)] -> Exp (a, b, c, d, e)

Num a => Monoid (Exp (Sum a))

Instance details

Methods

mempty :: Exp (Sum a)

mappend :: Exp (Sum a) -> Exp (Sum a) -> Exp (Sum a)

mconcat :: [Exp (Sum a)] -> Exp (Sum a)

Num a => Monoid (Exp (Product a))

Instance details

Methods

mempty :: Exp (Product a)

mappend :: Exp (Product a) -> Exp (Product a) -> Exp (Product a)

mconcat :: [Exp (Product a)] -> Exp (Product a)

Elt a => Elt (Sum a)

Instance details

Methods

eltType :: Sum a -> TupleType (EltRepr (Sum a))

fromElt :: Sum a -> EltRepr (Sum a)

toElt :: EltRepr (Sum a) -> Sum a

Elt a => Elt (Product a)

Instance details

Methods

eltType :: Product a -> TupleType (EltRepr (Product a))

fromElt :: Product a -> EltRepr (Product a)

toElt :: EltRepr (Product a) -> Product a

Eq a => Eq (Sum a)

Instance details

Methods

(==) :: Exp (Sum a) -> Exp (Sum a) -> Exp Bool

(/=) :: Exp (Sum a) -> Exp (Sum a) -> Exp Bool

Eq a => Eq (Product a)

Instance details

Methods

(==) :: Exp (Product a) -> Exp (Product a) -> Exp Bool

(/=) :: Exp (Product a) -> Exp (Product a) -> Exp Bool

Ord a => Ord (Sum a)

Instance details

Methods

(<) :: Exp (Sum a) -> Exp (Sum a) -> Exp Bool

(>) :: Exp (Sum a) -> Exp (Sum a) -> Exp Bool

(<=) :: Exp (Sum a) -> Exp (Sum a) -> Exp Bool

(>=) :: Exp (Sum a) -> Exp (Sum a) -> Exp Bool

min :: Exp (Sum a) -> Exp (Sum a) -> Exp (Sum a)

max :: Exp (Sum a) -> Exp (Sum a) -> Exp (Sum a)

compare :: Exp (Sum a) -> Exp (Sum a) -> Exp Ordering

Ord a => Ord (Product a)

Instance details

Methods

(<) :: Exp (Product a) -> Exp (Product a) -> Exp Bool

(>) :: Exp (Product a) -> Exp (Product a) -> Exp Bool

(<=) :: Exp (Product a) -> Exp (Product a) -> Exp Bool

(>=) :: Exp (Product a) -> Exp (Product a) -> Exp Bool

min :: Exp (Product a) -> Exp (Product a) -> Exp (Product a)

max :: Exp (Product a) -> Exp (Product a) -> Exp (Product a)

compare :: Exp (Product a) -> Exp (Product a) -> Exp Ordering