Data.Array.Accelerate.Data.Semigroup

Description

Semigroup instances for Accelerate

Since: 1.2.0.0

Documentation

class Semigroup a where

The class of semigroups (types with an associative binary operation).

Since: 4.9.0.0

Methods

(<>) :: a -> a -> a infixr 6

An associative operation.

(a <> b) <> c = a <> (b <> c)

If a is also a Monoid we further require

(<>) = mappend

sconcat :: NonEmpty a -> a

Reduce a non-empty list with <>

The default definition should be sufficient, but this can be overridden for efficiency.

stimes :: Integral b => b -> a -> a

Repeat a value n times.

Given that this works on a Semigroup it is allowed to fail if you request 0 or fewer repetitions, and the default definition will do so.

By making this a member of the class, idempotent semigroups and monoids can upgrade this to execute in O(1) by picking stimes = stimesIdempotent or stimes = stimesIdempotentMonoid respectively.

newtype Min a :: * -> *

Constructors

Min

Fields

getMin :: a

newtype Max a :: * -> *

Constructors

Max

Fields

getMax :: a

Orphan instances

Elt a => Unlift Exp (Min (Exp a))

Methods

unlift :: Exp (Plain (Min (Exp a))) -> Min (Exp a)

Elt a => Unlift Exp (Max (Exp a))

Methods

unlift :: Exp (Plain (Max (Exp a))) -> Max (Exp a)

(Lift Exp a, Elt (Plain a)) => Lift Exp (Min a)

Associated Types

type Plain (Min a) :: *

Methods

lift :: Min a -> Exp (Plain (Min a))

(Lift Exp a, Elt (Plain a)) => Lift Exp (Max a)

Associated Types

type Plain (Max a) :: *

Methods

lift :: Max a -> Exp (Plain (Max a))

Bounded a => Bounded (Exp (Min a))

Methods

minBound :: Exp (Min a)

maxBound :: Exp (Min a)

Bounded a => Bounded (Exp (Max a))

Methods

minBound :: Exp (Max a)

maxBound :: Exp (Max a)

Num a => Num (Exp (Min a))

Methods

(+) :: Exp (Min a) -> Exp (Min a) -> Exp (Min a)

(-) :: Exp (Min a) -> Exp (Min a) -> Exp (Min a)

(*) :: Exp (Min a) -> Exp (Min a) -> Exp (Min a)

negate :: Exp (Min a) -> Exp (Min a)

abs :: Exp (Min a) -> Exp (Min a)

signum :: Exp (Min a) -> Exp (Min a)

fromInteger :: Integer -> Exp (Min a)

Num a => Num (Exp (Max a))

Methods

(+) :: Exp (Max a) -> Exp (Max a) -> Exp (Max a)

(-) :: Exp (Max a) -> Exp (Max a) -> Exp (Max a)

(*) :: Exp (Max a) -> Exp (Max a) -> Exp (Max a)

negate :: Exp (Max a) -> Exp (Max a)

abs :: Exp (Max a) -> Exp (Max a)

signum :: Exp (Max a) -> Exp (Max a)

fromInteger :: Integer -> Exp (Max a)

Semigroup (Exp ())

Methods

(<>) :: Exp () -> Exp () -> Exp ()

sconcat :: NonEmpty (Exp ()) -> Exp ()

stimes :: Integral b => b -> Exp () -> Exp ()

(Elt a, Elt b, Semigroup (Exp a), Semigroup (Exp b)) => Semigroup (Exp (a, b))

Methods

(<>) :: Exp (a, b) -> Exp (a, b) -> Exp (a, b)

sconcat :: NonEmpty (Exp (a, b)) -> Exp (a, b)

stimes :: Integral b => b -> Exp (a, b) -> Exp (a, b)

(Elt a, Elt b, Elt c, Semigroup (Exp a), Semigroup (Exp b), Semigroup (Exp c)) => Semigroup (Exp (a, b, c))

Methods

(<>) :: Exp (a, b, c) -> Exp (a, b, c) -> Exp (a, b, c)

sconcat :: NonEmpty (Exp (a, b, c)) -> Exp (a, b, c)

stimes :: Integral b => b -> Exp (a, b, c) -> Exp (a, b, c)

(Elt a, Elt b, Elt c, Elt d, Semigroup (Exp a), Semigroup (Exp b), Semigroup (Exp c), Semigroup (Exp d)) => Semigroup (Exp (a, b, c, d))

Methods

(<>) :: Exp (a, b, c, d) -> Exp (a, b, c, d) -> Exp (a, b, c, d)

sconcat :: NonEmpty (Exp (a, b, c, d)) -> Exp (a, b, c, d)

stimes :: Integral b => b -> Exp (a, b, c, d) -> Exp (a, b, c, d)

(Elt a, Elt b, Elt c, Elt d, Elt e, Semigroup (Exp a), Semigroup (Exp b), Semigroup (Exp c), Semigroup (Exp d), Semigroup (Exp e)) => Semigroup (Exp (a, b, c, d, e))

Methods

(<>) :: Exp (a, b, c, d, e) -> Exp (a, b, c, d, e) -> Exp (a, b, c, d, e)

sconcat :: NonEmpty (Exp (a, b, c, d, e)) -> Exp (a, b, c, d, e)

stimes :: Integral b => b -> Exp (a, b, c, d, e) -> Exp (a, b, c, d, e)

Ord a => Semigroup (Exp (Min a))

Methods

(<>) :: Exp (Min a) -> Exp (Min a) -> Exp (Min a)

sconcat :: NonEmpty (Exp (Min a)) -> Exp (Min a)

stimes :: Integral b => b -> Exp (Min a) -> Exp (Min a)

Ord a => Semigroup (Exp (Max a))

Methods

(<>) :: Exp (Max a) -> Exp (Max a) -> Exp (Max a)

sconcat :: NonEmpty (Exp (Max a)) -> Exp (Max a)

stimes :: Integral b => b -> Exp (Max a) -> Exp (Max a)

(Ord a, Bounded a) => Monoid (Exp (Min a))

Methods

mempty :: Exp (Min a)

mappend :: Exp (Min a) -> Exp (Min a) -> Exp (Min a)

mconcat :: [Exp (Min a)] -> Exp (Min a)

(Ord a, Bounded a) => Monoid (Exp (Max a))

Methods

mempty :: Exp (Max a)

mappend :: Exp (Max a) -> Exp (Max a) -> Exp (Max a)

mconcat :: [Exp (Max a)] -> Exp (Max a)

Elt a => Elt (Min a)

Methods

eltType :: Min a -> TupleType (EltRepr (Min a))

fromElt :: Min a -> EltRepr (Min a)

toElt :: EltRepr (Min a) -> Min a

Elt a => Elt (Max a)

Methods

eltType :: Max a -> TupleType (EltRepr (Max a))

fromElt :: Max a -> EltRepr (Max a)

toElt :: EltRepr (Max a) -> Max a

Eq a => Eq (Min a)

Methods

(==) :: Exp (Min a) -> Exp (Min a) -> Exp Bool

(/=) :: Exp (Min a) -> Exp (Min a) -> Exp Bool

Eq a => Eq (Max a)

Methods

(==) :: Exp (Max a) -> Exp (Max a) -> Exp Bool

(/=) :: Exp (Max a) -> Exp (Max a) -> Exp Bool

Ord a => Ord (Min a)

Methods

(<) :: Exp (Min a) -> Exp (Min a) -> Exp Bool

(>) :: Exp (Min a) -> Exp (Min a) -> Exp Bool

(<=) :: Exp (Min a) -> Exp (Min a) -> Exp Bool

(>=) :: Exp (Min a) -> Exp (Min a) -> Exp Bool

min :: Exp (Min a) -> Exp (Min a) -> Exp (Min a)

max :: Exp (Min a) -> Exp (Min a) -> Exp (Min a)

compare :: Exp (Min a) -> Exp (Min a) -> Exp Ordering

Ord a => Ord (Max a)

Methods

(<) :: Exp (Max a) -> Exp (Max a) -> Exp Bool

(>) :: Exp (Max a) -> Exp (Max a) -> Exp Bool

(<=) :: Exp (Max a) -> Exp (Max a) -> Exp Bool

(>=) :: Exp (Max a) -> Exp (Max a) -> Exp Bool

min :: Exp (Max a) -> Exp (Max a) -> Exp (Max a)

max :: Exp (Max a) -> Exp (Max a) -> Exp (Max a)

compare :: Exp (Max a) -> Exp (Max a) -> Exp Ordering