Data.Array.Accelerate.Interpreter

Description

This interpreter is meant to be a reference implementation of the semantics of the embedded array language. The emphasis is on defining the semantics clearly, not on performance.

Interpret an array expression

class (Typeable a, Typeable (ArrRepr a)) => Arrays a

Arrays consists of nested tuples of individual Arrays, currently up to 15-elements wide. Accelerate computations can thereby return multiple results.

Minimal complete definition

arrays, flavour, toArr, fromArr

run :: Arrays a => Acc a -> a

Run a complete embedded array program using the reference interpreter.

run1 :: (Arrays a, Arrays b) => (Acc a -> Acc b) -> a -> b

Prepare and run an embedded array program of one argument